Mathematica code to compute tanh solutions of nonlinear differential-difference equations (DDEs)
A Mathematica package for the symbolic computation of tanh solutions of polynomial systems of nonlinear differential-difference equations (DDEs). The code was written in 2001 and last updated in 2010.

The 2001 version of the program (with manual) was added to the Computer Physics Communications Program Library, initially hosted at Queen’s University of Belfast (North Ireland) and now maintained by Mendeley Data. Search for DDESpecialSolutions.m.

Written by Douglas Baldwin, Ünal Göktas, and Willy Hereman.